test:basic unit test for cri manager
